+++ NEWS.md
@@ -1,3 +1,26 @@
+# wdman 0.1.5
+
+* Use binman::sem_ver for versioing.
+* Remove the v in gecko versions.
+
+# wdman 0.1.4
+
+* Added tests for driver functions.
+
+# wdman 0.1.3
+
+* Added selenium function.
+* Added iedriver function.
+
+# wdman 0.1.2
+
+* Added phantomjs function.
+
+# wdman 0.1.1
+
+* Added chrome function.
+* Added gecko function.
+
 # wdman 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.